requirements.txt
pyrogram==2.0.97
TgCrypto==1.2.5
Pillow>=8.4.0
pytz>=2021.3
PyYAML>=6.0
coloredlogs>=15.0.1
psutil>=5.8.0
httpx~=0.23.3
apscheduler~=3.9.1.post1
sqlitedict~=2.1.0
casbin==1.17.5
sentry-sdk==1.14.0
PyQRCode>=1.2.1
PyPng
fastapi~=0.89.1
amis-python
python-jose
uvicorn
pydantic~=1.10.4
starlette~=0.22.0
